pyproject.toml
[project]
name = "intern-report-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-docx>=1.2.0",
]
